+++ src/constants/configs.ts
@@ -3,10 +3,6 @@
 
 export const LOCAL_HOST: string = '127.0.0.1';
 
-export const SAVE_ALL_BEFORE_LAUNCH_SETTING_KEY: string = 'java.test.saveAllBeforeLaunchTest';
-
-export const BUILD_BEFORE_LAUNCH_SETTING_KEY: string = 'java.test.forceBuildBeforeLaunchTest';
-
 export const LOG_FILE_NAME: string = 'java_test_runner.log';
 export const LOG_FILE_MAX_SIZE: number = 5 * 1024 * 1024;
 export const LOG_FILE_MAX_NUMBER: number = 2;